rename coef_function.h/cpp
