+++ source/core/calibration/CMakeLists.txt
@@ -2,7 +2,7 @@ set(dir ${CMAKE_CURRENT_SOURCE_DIR})
 
 set(SOURCES
   ${dir}/parameter.cpp
-  ${dir}/coef_function.cpp
+  ${dir}/CoefFunction.cpp
   ${dir}/CoefFunctionFactory.cpp
   ${dir}/Polynomial.cpp
   ${dir}/Calibration.cpp
@@ -10,7 +10,7 @@ set(SOURCES
 
 set(HEADERS
   ${dir}/parameter.h
-  ${dir}/coef_function.h
+  ${dir}/CoefFunction.h
   ${dir}/CoefFunctionFactory.h
   ${dir}/Polynomial.h
   ${dir}/Calibration.h
