+++ source/core/calibration/CoefFunction.cpp
@@ -1,4 +1,4 @@
-#include <core/calibration/coef_function.h>
+#include <core/calibration/CoefFunction.h>
 
 namespace DAQuiri
 {
