+++ tests/core/calibration/CMakeLists.txt
@@ -2,7 +2,7 @@ set(dir ${CMAKE_CURRENT_SOURCE_DIR})
 
 set(SOURCES
   ${dir}/parameter.cpp
-  ${dir}/coef_function.cpp
+  ${dir}/CoefFunctionTest.cpp
   ${dir}/CoefFunctionFactoryTest.cpp
   ${dir}/PolynomialTest.cpp
   ${dir}/CalibrationTest.cpp
